requirements.txt
requests
pandas
sqlalchemy
pyodbc
python-dotenv